+++ RuneTerra Server/src/com/runeterra/world/entity/combat/method/impl/specials/ArmadylCrossbowCombatMethod.java
@@ -56,7 +56,7 @@ public class ArmadylCrossbowCombatMethod implements CombatMethod {
 	}
 
 	@Override
-	public int getAttackSpeed(Character character) {
+	public float getAttackSpeed(Character character) {
 		return character.getBaseAttackSpeed();
 	}
