+++ RuneTerra Server/src/com/runeterra/world/entity/combat/method/impl/specials/BandosGodswordCombatMethod.java
@@ -38,7 +38,7 @@ public class BandosGodswordCombatMethod implements CombatMethod {
 	}
 
 	@Override
-	public int getAttackSpeed(Character character) {
+	public float getAttackSpeed(Character character) {
 		return character.getBaseAttackSpeed();
 	}
