+++ RuneTerra Server/src/com/runeterra/world/entity/combat/method/impl/specials/ZamorakGodswordCombatMethod.java
@@ -36,7 +36,7 @@ public class ZamorakGodswordCombatMethod implements CombatMethod {
 	}
 
 	@Override
-	public int getAttackSpeed(Character character) {
+	public float getAttackSpeed(Character character) {
 		return character.getBaseAttackSpeed();
 	}
