+++ RuneTerra Server/src/com/runeterra/world/entity/impl/player/Player.java
@@ -132,13 +132,13 @@ public class Player extends Character {
 	}
 
 	@Override
-	public int getBaseAttackSpeed() {
+	public float getBaseAttackSpeed() {
 
 		// Gets attack speed for player's weapon
 		// If player is using magic, attack speed is
 		// Calculated in the MagicCombatMethod class.
 
-		int speed = getCombat().getWeapon().getSpeed();
+		float speed = getCombat().getWeapon().getSpeed();
 		// if(getCombat().getFightType().getStyle() == FightStyle.AGGRESSIVE) {
 		if (getCombat().getFightType().getStyle().toString().toLowerCase().contains("rapid")) {
 			speed--;
